refatorado componente Post, componentizada funcao que gera o texto da tooltip e subcomponente PostContent
